# bark-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "barkd REST API",
    "description": "A simple REST API for barkd, a wallet daemon for integrating bitcoin payments into your app over HTTP. Supports self-custodial Lightning, Ark, and on-chain out of the box.\n\nbarkd is a long-running daemon best suited for always-on or high-connectivity environments like nodes, servers, desktops, and point-of-sale terminals.\n\nAll endpoints return JSON. Amounts are denominated in satoshis.",
    "contact": {
      "name": "Second Team",
      "email": "hello@second.tech"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.7.1"
  },
  "paths": {
...
    "/api/v1/fees/board": {
      "get": {
        "tags": [
          "fees"
        ],
        "summary": "Estimate board fee",
        "description": "Estimates the Ark protocol fee for boarding the specified amount of on-chain bitcoin. The net amount is what the user receives as a VTXO. Does not include the on-chain transaction fee for the board anchor transaction.",
        "operationId": "board_fee",
        "parameters": [
          {
            "name": "amount_sat",
            "in": "query",
            "description": "The amount in satoshis to board",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the fee estimate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid amount",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/lightning/pay": {
      "get": {
        "tags": [
          "fees"
        ],
        "summary": "Estimate Lightning send fee",
        "description": "Estimates the fee for sending the specified amount over Lightning. The net amount is what the recipient receives. The fee depends on the VTXOs selected and their expiry, and includes funded recovery reserves. Uses the same transaction-builder validation as payment submission. Returns an error when no usable input combination is found.",
        "operationId": "lightning_send_fee",
        "parameters": [
          {
            "name": "amount_sat",
            "in": "query",
            "description": "The amount in satoshis to send over Lightning",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the fee estimate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid amount",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/lightning/receive": {
      "get": {
        "tags": [
          "fees"
        ],
        "summary": "Estimate Lightning receive fee",
        "description": "Estimates the fee for receiving the specified amount over Lightning. The gross amount is the Lightning payment amount, and the net amount is what the user receives as a VTXO after the Ark server deducts its fee.",
        "operationId": "lightning_receive_fee",
        "parameters": [
          {
            "name": "amount_sat",
            "in": "query",
            "description": "The amount in satoshis to receive over Lightning",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the fee estimate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid amount",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/offboard": {
      "post": {
        "tags": [
          "fees"
        ],
        "summary": "Estimate offboard fee",
        "description": "Estimates the fee for offboarding a specific set of VTXOs to the given on-chain address. Each VTXO is offboarded in full. The gross amount is the total value of the selected VTXOs, and the net amount is what the user receives on-chain after fees. The fee depends on the destination address type, current fee rates, and VTXO expiry.",
        "operationId": "offboard_fee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OffboardFeeEstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns the fee estimate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "No VTXO IDs provided, or an invalid VTXO id or address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "404": {
            "description": "One of the VTXOs wasn't found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/offboard-all": {
      "get": {
        "tags": [
          "fees"
        ],
        "summary": "Estimate offboard-all fee",
        "description": "Estimates the fee for offboarding the entire Ark balance to the given on-chain address. The gross amount is the total spendable balance, and the net amount is what the user receives on-chain after fees. The fee depends on the destination address type, current fee rates, and VTXO expiry.",
        "operationId": "offboard_all_fee",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "description": "The destination Bitcoin address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the fee estimate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/onchain": {
      "get": {
        "tags": [
          "fees"
        ],
        "summary": "Get on-chain fee rates",
        "description": "Returns the current mempool fee rates from the chain source at three confirmation targets: fast (~1 block), regular (~3 blocks), and slow (~6 blocks). Rates are in sat/vB, rounded up.",
        "operationId": "onchain_fee_rates",
        "responses": {
          "200": {
            "description": "Returns current mempool fee rates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnchainFeeRatesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/send-onchain": {
      "get": {
        "tags": [
          "fees"
        ],
        "summary": "Estimate send-onchain fee",
        "description": "Estimates the total fee for sending bitcoin from the Ark wallet to an on-chain address. The fee depends on the destination address type and current fee rates. The gross amount is what the user pays (including VTXOs spent), and the net amount is what the recipient receives on-chain.",
        "operationId": "send_onchain_fee",
        "parameters": [
          {
            "name": "amount_sat",
            "in": "query",
            "description": "The amount in satoshis to send on-chain",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "address",
            "in": "query",
            "description": "The destination Bitcoin address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the fee estimate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeEstimateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid amount or address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/lightning/pay": {
      "post": {
        "tags": [
          "lightning"
        ],
        "summary": "Send a Lightning payment",
        "description": "Sends a payment to a Lightning destination. Accepts a BOLT11 invoice, BOLT12 offer, or Lightning address. The `amount_sat` field is required for Lightning addresses but optional for invoices and offers. Comments are only supported for Lightning addresses.",
        "operationId": "pay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LightningPayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Returns success message, optionally with preimage if payment was immediately settled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LightningPayResponse"
                }
              }
            }
          },
          "400": {
            "description": "The provided destination is not a valid BOLT11 invoice, BOLT12 offer, or Lightning address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/wallet/addresses/next": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Generate Ark address",
        "description": "Generates a new Ark receiving address. Each call returns the next unused address from the wallet's HD keychain.",
        "operationId": "address",
        "responses": {
          "200": {
            "description": "Returns the Ark address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArkAddressResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/wallet/balance": {
      "get": {
        "tags": [
          "wallet"
        ],
        "summary": "Get wallet balance",
        "description": "Returns the wallet balance broken down by category: spendable sats available for immediate use, sats pending in an Ark round, sats locked in outgoing or incoming Lightning payments, sats awaiting board confirmation, and sats in a pending exit. The balance is computed from local state, which the background daemon keeps reasonably fresh (Lightning syncs every second, mailbox and boards every 30 seconds). For the most up-to-date figures, call `sync` before this endpoint.",
        "operationId": "balance",
        "responses": {
          "200": {
            "description": "Returns the wallet balance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Balance"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/wallet/refresh/all": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Refresh all VTXOs",
        "description": "Registers all spendable VTXOs for refresh in the next Ark round. The input VTXOs are locked immediately and will be forfeited once the round completes, yielding new VTXOs with a fresh expiry. The background daemon automatically participates in the round and progresses it to completion. Use the `rounds` endpoint to track progress.",
        "operationId": "refresh_all",
        "responses": {
          "200": {
            "description": "Returns the refresh result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingRoundInfo"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadRequestError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "FeeEstimateResponse": {
        "type": "object",
        "description": "A fee estimate for an Ark wallet operation.",
        "required": [
          "gross_amount_sat",
          "fee_sat",
          "net_amount_sat",
          "vtxos_spent"
        ],
        "properties": {
          "fee_sat": {
            "type": "integer",
            "format": "int64",
            "description": "The fee portion (in satoshis)",
            "minimum": 0
          },
          "gross_amount_sat": {
            "type": "integer",
            "format": "int64",
            "description": "The total amount including fees (in satoshis)",
            "minimum": 0
          },
          "net_amount_sat": {
            "type": "integer",
            "format": "int64",
            "description": "The amount excluding fees (in satoshis). For sends, this is the amount\nthe recipient receives. For receives, this is the amount the user gets.",
            "minimum": 0
          },
          "vtxos_spent": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The VTXOs that would be spent for this operation"
          }
        }
      },
...
      "InternalServerError": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "NotFoundError": {
        "type": "object",
        "required": [
          "resources",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "resources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "OffboardFeeEstimateRequest": {
        "type": "object",
        "description": "Request body for estimating the fee of offboarding a specific set of VTXOs.",
        "required": [
          "address",
          "vtxos"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "The destination Bitcoin address. The fee depends on its script type."
          },
          "vtxos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The IDs of the VTXOs to offboard. Each is offboarded in full."
          }
        }
      },
...
      "OnchainFeeRatesResponse": {
        "type": "object",
        "description": "Mempool fee rates for on-chain transactions.",
        "required": [
          "fast_sat_per_vb",
          "regular_sat_per_vb",
          "slow_sat_per_vb"
        ],
        "properties": {
          "fast_sat_per_vb": {
            "type": "integer",
            "format": "int64",
            "description": "Fee rate targeting ~1 block confirmation (sat/vB)",
            "minimum": 0
          },
          "regular_sat_per_vb": {
            "type": "integer",
            "format": "int64",
            "description": "Fee rate targeting ~3 block confirmation (sat/vB)",
            "minimum": 0
          },
          "slow_sat_per_vb": {
            "type": "integer",
            "format": "int64",
            "description": "Fee rate targeting ~6 block confirmation (sat/vB)",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "AuthToken",
        "description": "Base64url-encoded auth token"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}